campus_trade/requirements.txt
# Core
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2

# Database
PyMySQL==1.1.1

# File Upload & Image Processing
Pillow==12.2.0

# Configuration
python-dotenv==1.0.1

# Form Validation
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
coverage==7.6.12

# Code Quality
ruff==0.11.0
black==25.1.0
